requirements.txt
# No external dependencies required
# Script uses only Python standard library modules 

# External dependencies
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0 